dom/contacts/tests/test_contacts_basics.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=674720
-->
<head>
  <title>Test for Bug 674720 WebContacts</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=674720">Mozilla Bug 674720</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

"use strict";

if (SpecialPowers.isMainProcess()) {
  SpecialPowers.Cu.import("resource://gre/modules/ContactService.jsm");
  SpecialPowers.Cu.import("resource://gre/modules/PermissionPromptHelper.jsm");
}

SpecialPowers.addPermission("contacts-write", true, document);
SpecialPowers.addPermission("contacts-read", true, document);
SpecialPowers.addPermission("contacts-create", true, document);

var isAndroid = (navigator.userAgent.indexOf("Android") !== -1);
var androidVersion = SpecialPowers.Cc['@mozilla.org/system-info;1']
                                  .getService(SpecialPowers.Ci.nsIPropertyBag2)
                                  .getProperty('version');

// For Sorting
var c1 = {
  name: ["a a"],
  familyName: ["a"],
  givenName: ["a"],
};

var c2 = {
  name: ["b b"],
  familyName: ["b"],
  givenName: ["b"],
};

var c3 = {
  name: ["c c", "a a", "b b"],
  familyName: ["c","a","b"],
  givenName: ["c","a","b"],
};

var c4 = {
  name: ["c c", "a a", "c c"],
  familyName: ["c","a","c"],
  givenName: ["c","a","c"],
};

var c5 = {
  familyName: [],
  givenName: [],
};

var c6 = {
  name: ["e"],
  familyName: ["e","e","e"],
  givenName: ["e","e","e"],
};

var c7 = {
  name: ["e"],
  familyName: ["e","e","e"],
  givenName: ["e","e","e"],
};

var c8 = {
  name: ["e"],
  familyName: ["e","e","e"],
  givenName: ["e","e","e"],
};

var adr1 = {
  type: ["work"],
  streetAddress: "street 1",
  locality: "locality 1",
  region: "region 1",
  postalCode: "postal code 1",
  countryName: "country 1"
};

var adr2 = {
  type: ["home, fax"],
  streetAddress: "street2",
  locality: "locality2",
  region: "region2",
  postalCode: "postal code2",
  countryName: "country2"
};

var properties1 = {
  // please keep capital letters at the start of these names
  name: ["Test1 TestFamilyName", "Test2 Wagner"],
  familyName: ["TestFamilyName","Wagner"],
  givenName: ["Test1","Test2"],
  nickname: ["nicktest"],
  tel: [{type: ["work"], value: "123456", carrier: "testCarrier"} , {type: ["home", "fax"], value: "+55 (31) 9876-3456"}, {type: ["home"], value: "+49 451 491934"}],
  adr: [adr1],
  email: [{type: ["work"], value: "x@y.com"}],
};

var properties2 = {
  name: ["dummyHonorificPrefix dummyGivenName dummyFamilyName dummyHonorificSuffix", "dummyHonorificPrefix2"],
  familyName: ["dummyFamilyName"],
  givenName: ["dummyGivenName"],
  honorificPrefix: ["dummyHonorificPrefix","dummyHonorificPrefix2"],
  honorificSuffix: ["dummyHonorificSuffix"],
  additionalName: ["dummyadditionalName"],
  nickname: ["dummyNickname"],
  tel: [{type: ["test"], value: "7932012345", carrier: "myCarrier", pref: 1},{type: ["home", "custom"], value: "7932012346", pref: 0}],
  email: [{type: ["test"], value: "a@b.c"}, {value: "b@c.d", pref: 1}],
  adr: [adr1, adr2],
  impp: [{type: ["aim"], value:"im1", pref: 1}, {value: "im2"}],
  org: ["org1", "org2"],
  jobTitle: ["boss", "superboss"],
  note: ["test note"],
  category: ["cat1", "cat2"],
  url: [{type: ["work", "work2"], value: "www.1.com", pref: 1}, {value:"www2.com"}],
  bday: new Date("1980, 12, 01"),
  anniversary: new Date("2000, 12, 01"),
  sex: "male",
  genderIdentity: "test",
  key: ["ERPJ394GJJWEVJ0349GJ09W3H4FG0WFW80VHW3408GH30WGH348G3H"]
};

var sample_id1;
var sample_id2;

var createResult1;
var createResult2;

var findResult1;
var findResult2;

function clearTemps() {
  sample_id1 = null;
  sample_id2 = null;
  createResult1 = null;
  createResult2 = null;
  findResult1 = null;
  findResult2 = null;
}

function onUnwantedSuccess() {
  ok(false, "onUnwantedSuccess: shouldn't get here");
}

function onFailure() {
  ok(false, "in on Failure!");
  next();
}

function checkStr(str1, str2, msg) {
  if (str1 ^ str2) {
    ok(false, "Expected both strings to be either present or absent");
    return;
  }
  if (!str1 || str1 == "null") {
    str1 = null;
  }
  if (!str2 || str2 == "null") {
    str2 = null;
  }
  is(str1, str2, msg);
}

function checkStrArray(str1, str2, msg) {
  function normalize_falsy(v) {
    if (!v || v == "null" || v == "undefined") {
      return "";
    }
    return v;
  }
  function optArray(val) {
    return Array.isArray(val) ? val : [val];
  }
  str1 = optArray(str1).map(normalize_falsy).filter(v => v != "");
  str2 = optArray(str2).map(normalize_falsy).filter(v => v != "");
  ise(JSON.stringify(str1), JSON.stringify(str2), msg);
}

function checkPref(pref1, pref2) {
  // If on Android treat one preference as 0 and the other as undefined as matching
  if (isAndroid) {
    if ((!pref1 && pref2 == undefined) || (pref1 == undefined && !pref2)) {
      pref1 = false;
      pref2 = false;
    }
  }
  ise(!!pref1, !!pref2, "Same pref");
}

function checkAddress(adr1, adr2) {
  if (adr1 ^ adr2) {
    ok(false, "Expected both adrs to be either present or absent");
    return;
  }
  checkStrArray(adr1.type, adr2.type, "Same type");
  checkStr(adr1.streetAddress, adr2.streetAddress, "Same streetAddress");
  checkStr(adr1.locality, adr2.locality, "Same locality");
  checkStr(adr1.region, adr2.region, "Same region");
  checkStr(adr1.postalCode, adr2.postalCode, "Same postalCode");
  checkStr(adr1.countryName, adr2.countryName, "Same countryName");
  checkPref(adr1.pref, adr2.pref);
}

function checkField(field1, field2) {
  if (field1 ^ field2) {
    ok(false, "Expected both fields to be either present or absent");
    return;
  }
  checkStrArray(field1.type, field2.type, "Same type");
  checkStr(field1.value, field2.value, "Same value");
  checkPref(field1.pref, field2.pref);
}

function checkTel(tel1, tel2) {
  if (tel1 ^ tel2) {
    ok(false, "Expected both tels to be either present or absent");
    return;
  }
  checkField(tel1, tel2);
  checkStr(tel1.carrier, tel2.carrier, "Same carrier");
}

function checkCategory(category1, category2) {
  // Android adds contacts to the a default category. This should be removed from the
  // results before comparing them
  if (isAndroid) {
    category1 = removeAndroidDefaultCategory(category1);
    category2 = removeAndroidDefaultCategory(category2);
  }
  checkStrArray(category1, category2, "Same Category")
}

function removeAndroidDefaultCategory(category) {
  if (!category) {
    return category;
  }

  var result = [];

  for (var i of category) {
    // Some devices may return the full group name (prefixed with "System Group: ")
    if (i != "My Contacts" && i != "System Group: My Contacts") {
      result.push(i);
    }
  }

  return result;
}

function checkArrayField(array1, array2, func, msg) {
  if (!!array1 ^ !!array2) {
    ok(false, "Expected both arrays to be either present or absent: " + JSON.stringify(array1) + " vs. " + JSON.stringify(array2) + ". (" + msg + ")");
    return;
  }
  if (!array1 && !array2)  {
    ok(true, msg);
    return;
  }
  ise(array1.length, array2.length, "Same length");
  for (var i = 0; i < array1.length; ++i) {
    func(array1[i], array2[i], msg);
  }
}

function checkContacts(contact1, contact2) {
  if (!!contact1 ^ !!contact2) {
    ok(false, "Expected both contacts to be either present or absent");
    return;
  }
  checkStrArray(contact1.name, contact2.name, "Same name");
  checkStrArray(contact1.honorificPrefix, contact2.honorificPrefix, "Same honorificPrefix");
  checkStrArray(contact1.givenName, contact2.givenName, "Same givenName");
  checkStrArray(contact1.additionalName, contact2.additionalName, "Same additionalName");
  checkStrArray(contact1.familyName, contact2.familyName, "Same familyName");
  checkStrArray(contact1.honorificSuffix, contact2.honorificSuffix, "Same honorificSuffix");
  checkStrArray(contact1.nickname, contact2.nickname, "Same nickname");
  checkCategory(contact1.category, contact2.category);
  checkStrArray(contact1.org, contact2.org, "Same org");
  checkStrArray(contact1.jobTitle, contact2.jobTitle, "Same jobTitle");
  is(contact1.bday ? contact1.bday.valueOf() : null, contact2.bday ? contact2.bday.valueOf() : null, "Same birthday");
  checkStrArray(contact1.note, contact2.note, "Same note");
  is(contact1.anniversary ? contact1.anniversary.valueOf() : null , contact2.anniversary ? contact2.anniversary.valueOf() : null, "Same anniversary");
  checkStr(contact1.sex, contact2.sex, "Same sex");
  checkStr(contact1.genderIdentity, contact2.genderIdentity, "Same genderIdentity");
  checkStrArray(contact1.key, contact2.key, "Same key");

  checkArrayField(contact1.adr, contact2.adr, checkAddress, "Same adr");
  checkArrayField(contact1.tel, contact2.tel, checkTel, "Same tel");
  checkArrayField(contact1.email, contact2.email, checkField, "Same email");
  checkArrayField(contact1.url, contact2.url, checkField, "Same url");
  checkArrayField(contact1.impp, contact2.impp, checkField, "Same impp");
}

var req;
var index = 0;

var initialRev;

var defaultOptions = {
  sortBy: "givenName",
};

function checkRevision(revision, msg, then) {
  var revReq = mozContacts.getRevision();
  revReq.onsuccess = function(e) {
    is(e.target.result, initialRev+revision, msg);
    then();
  };
  // The revision function isn't supported on Android so treat on failure as success
  if (isAndroid) {
    revReq.onerror = function(e) {
      then();
    };
  } else {
    revReq.onerror = onFailure;
  }
}

function checkCount(count, msg, then) {
  var request = navigator.mozContacts.getCount();
  request.onsuccess = function(e) {
    is(e.target.result, count, msg);
    then();
  };
  request.onerror = onFailure;
}

var mozContacts = window.navigator.mozContacts;
var steps = [
  function() {
    req = mozContacts.getRevision();
    req.onsuccess = function(e) {
      initialRev = e.target.result;
      next();
    };

    // Android does not support the revision function. Treat errors as success.
    if (isAndroid) {
      req.onerror = function(e) {
        initialRev = 0;
        next();
      };
    } else {
      req.onerror = onFailure;
    }
  },
  function () {
    ok(true, "Deleting database");
    checkRevision(0, "Initial revision is 0", function() {
      req = mozContacts.clear();
      req.onsuccess = function () {
        ok(true, "Deleted the database");
        checkCount(0, "No contacts after clear", function() {
          checkRevision(1, "Revision was incremented on clear", next);
        });
      };
      req.onerror = onFailure;
    });
  },
  function () {
    ok(true, "Retrieving all contacts");
    req = mozContacts.find(defaultOptions);
    req.onsuccess = function () {
      is(req.result.length, 0, "Empty database.");
      checkRevision(1, "Revision was not incremented on find", next);
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding empty contact");
    createResult1 = new mozContact({});
    req = navigator.mozContacts.save(createResult1);
    req.onsuccess = function () {
      ok(createResult1.id, "The contact now has an ID.");
      sample_id1 = createResult1.id;
      checkCount(1, "1 contact after adding empty contact", function() {
        checkRevision(2, "Revision was incremented on save", next);
      });
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts");
    req = mozContacts.find(defaultOptions);
    req.onsuccess = function () {
      is(req.result.length, 1, "One contact.");
      findResult1 = req.result[0];
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Deleting empty contact");
    req = navigator.mozContacts.remove(findResult1);
    req.onsuccess = function () {
      var req2 = mozContacts.find(defaultOptions);
      req2.onsuccess = function () {
        is(req2.result.length, 0, "Empty Database.");
        clearTemps();
        checkRevision(3, "Revision was incremented on remove", next);
      }
      req2.onerror = onFailure;
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding a new contact1");
    createResult1 = new mozContact(properties1);

    mozContacts.oncontactchange = function(event) {
      is(event.contactID, createResult1.id, "Same contactID");
      is(event.reason, "create", "Same reason");
      next();
    }

    req = navigator.mozContacts.save(createResult1);
    req.onsuccess = function () {
      ok(createResult1.id, "The contact now has an ID.");
      sample_id1 = createResult1.id;
      checkContacts(createResult1, properties1);
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring 1");
    var options = {filterBy: ["givenName"],
                   filterOp: "startsWith",
                   filterValue: properties1.givenName[1].substring(0,3)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      // Some manual testing. Testint the testfunctions
      // tel: [{type: ["work"], value: "123456", carrier: "testCarrier"} , {type: ["home", "fax"], value: "+55 (31) 9876-3456"}],
      is(findResult1.tel[0].carrier, "testCarrier", "Same Carrier");
      is(findResult1.tel[0].type, "work", "Same type");
      is(findResult1.tel[0].value, "123456", "Same Value");
      is(findResult1.tel[1].type[1], "fax", "Same type");
      is(findResult1.tel[1].value, "+55 (31) 9876-3456", "Same Value");

      is(findResult1.adr[0].countryName, "country 1", "Same country");

      // email: [{type: ["work"], value: "x@y.com"}]
      is(findResult1.email[0].type, "work", "Same Type");
      is(findResult1.email[0].value, "x@y.com", "Same Value");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Searching for exact email");
    var options = {filterBy: ["email"],
                   filterOp: "equals",
                   filterValue: properties1.email[0].value};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, createResult1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring and update");
    mozContacts.oncontactchange = function(event) {
       is(event.contactID, findResult1.id, "Same contactID");
       is(event.reason, "update", "Same reason");
     }
    var options = {filterBy: ["givenName"],
                   filterOp: "startsWith",
                   filterValue: properties1.givenName[0].substring(0,3)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      findResult1.jobTitle = ["new Job"];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding a new contact");
    mozContacts.oncontactchange = function(event) {
       is(event.contactID, createResult2.id, "Same contactID");
       is(event.reason, "create", "Same reason");
     }
    createResult2 = new mozContact({name: ["newName"]});
    req = navigator.mozContacts.save(createResult2);
    req.onsuccess = function () {
      ok(createResult2.id, "The contact now has an ID.");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring 2");
    var options = {filterBy: ["givenName"],
                   filterOp: "startsWith",
                   filterValue: properties1.givenName[0].substring(0,3)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      checkContacts(createResult1, findResult1);
      next();
    };
    req.onerror = onFailure;
  },
  function() {
    ok(true, "Retrieving by name equality 1");
    var options = {filterBy: ["name"],
                   filterOp: "equals",
                   filterValue: properties1.name[0]};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      checkContacts(createResult1, findResult1);
      next();
    };
    req.onerror = onFailure;
  },
  function() {
    ok(true, "Retrieving by name equality 2");
    var options = {filterBy: ["name"],
                   filterOp: "equals",
                   filterValue: properties1.name[1]};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      checkContacts(createResult1, findResult1);
      next();
    };
    req.onerror = onFailure;
  },
  function() {
    ok(true, "Retrieving by name substring 1");
    var options = {filterBy: ["name"],
                   filterOp: "startsWith",
                   filterValue: properties1.name[0].substring(0,3).toLowerCase()};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      checkContacts(createResult1, findResult1);
      next();
    };
    req.onerror = onFailure;
  },
  function() {
    ok(true, "Retrieving by name substring 2");
    var options = {filterBy: ["name"],
                   filterOp: "startsWith",
                   filterValue: properties1.name[1].substring(0,3).toLowerCase()};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      checkContacts(createResult1, findResult1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Remove contact1");
    mozContacts.oncontactchange = function(event) {
      is(event.contactID, createResult1.id, "Same contactID");
      is(event.reason, "remove", "Same reason");
    }
    req = navigator.mozContacts.remove(createResult1);
    req.onsuccess = function () {
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring 3");
    var options = {filterBy: ["givenName"],
                   filterOp: "startsWith",
                   filterValue: properties1.givenName[1].substring(0,3)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 0, "Found no contact.");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Remove contact2");
    mozContacts.oncontactchange = function(event) {
      is(event.contactID, createResult2.id, "Same contactID");
      is(event.reason, "remove", "Same reason");
    }
    req = navigator.mozContacts.remove(createResult2);
    req.onsuccess = function () {
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring 4");
    var options = {filterBy: ["givenName"],
                   filterOp: "startsWith",
                   filterValue: properties1.givenName[1].substring(0,3)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 0, "Found no contact.");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Deleting database");
    mozContacts.oncontactchange = function(event) {
      is(event.contactID, "undefined", "Same contactID");
      is(event.reason, "remove", "Same reason");
    }
    req = mozContacts.clear();
    req.onsuccess = function () {
      ok(true, "Deleted the database");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Adding a new contact with properties1");
    createResult1 = new mozContact(properties1);
    mozContacts.oncontactchange = null;
    req = navigator.mozContacts.save(createResult1);
    req.onsuccess = function () {
      ok(createResult1.id, "The contact now has an ID.");
      sample_id1 = createResult1.id;
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring tel1");
    var options = {filterBy: ["tel"],
                   filterOp: "contains",
                   filterValue: properties1.tel[1].value.substring(2,5)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by tel exact");
    var options = {filterBy: ["tel"],
                   filterOp: "equals",
                   filterValue: "+55 319 8 7 6 3456"};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by tel exact with substring");
    var options = {filterBy: ["tel"],
                   filterOp: "equals",
                   filterValue: "3456"};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 0, "Found no contacts.");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by tel exact with substring");
    var options = {filterBy: ["tel"],
                   filterOp: "equals",
                   filterValue: "+55 (31)"};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 0, "Found no contacts.");
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by tel match national number");
    var options = {filterBy: ["tel"],
                   filterOp: "match",
                   filterValue: "3198763456"};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by tel match national format");
    var options = {filterBy: ["tel"],
                   filterOp: "match",
                   filterValue: "0451 491934"};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by tel match entered number");
    var options = {filterBy: ["tel"],
                   filterOp: "match",
                   filterValue: "123456"};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by tel match international number");
    var options = {filterBy: ["tel"],
                   filterOp: "match",
                   filterValue: "+55 31 98763456"};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by match with field other than tel");
    var options = {filterBy: ["givenName"],
                   filterOp: "match",
                   filterValue: "my friends call me 555-4040"};
    req = mozContacts.find(options);
    req.onsuccess = onUnwantedSuccess;
    req.onerror = function() {
      ok(true, "Failed");
      next();
    }
  },
  function () {
    ok(true, "Retrieving by substring tel2");
    var options = {filterBy: ["tel"],
                   filterOp: "startsWith",
                   filterValue: "9876"};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring tel3");
    var options = {filterBy: ["tel"],
                   filterOp: "startsWith",
                   filterValue: "98763456"};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring 5");
    var options = {filterBy: ["givenName"],
                   filterOp: "startsWith",
                   filterValue: properties1.givenName[0].substring(0,3)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring 6");
    var options = {filterBy: ["familyName", "givenName"],
                   filterOp: "startsWith",
                   filterValue: properties1.givenName[0].substring(0,3)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving by substring3, Testing multi entry");
    var options = {filterBy: ["givenName", "familyName"],
                   filterOp: "startsWith",
                   filterValue: properties1.familyName[1].substring(0,3).toLowerCase()};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving all contacts");
    req = mozContacts.find(defaultOptions);
    req.onsuccess = function() {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(createResult1, findResult1);
      if (!isAndroid) {
        ok(findResult1.updated, "Has updated field");
        ok(findResult1.published, "Has published field");
      }
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Modifying contact1");
    if (!findResult1) {
      SpecialPowers.executeSoon(next);
    } else {
      findResult1.impp = properties1.impp = [{value:"phil impp"}];
      req = navigator.mozContacts.save(findResult1);
      req.onsuccess = function () {
        var req2 = mozContacts.find(defaultOptions);
        req2.onsuccess = function() {
          is(req2.result.length, 1, "Found exactly 1 contact.");
          findResult2 = req2.result[0];
          ok(findResult2.id == sample_id1, "Same ID");
          checkContacts(findResult2, properties1);
          is(findResult2.impp.length, 1, "Found exactly 1 IMS info.");
          next();
        };
        req2.onerror = onFailure;
      };
      req.onerror = onFailure;
    }
  },
  function() {
    // Android does not support published/updated fields. Skip this.
    if (isAndroid) {
      next();
      return;
    }

    ok(true, "Saving old contact, should abort!");
    req = mozContacts.save(createResult1);
    req.onsuccess = onUnwantedSuccess;
    req.onerror   = function() { ok(true, "Successfully declined updating old contact!"); next(); };
  },
  function () {
    ok(true, "Retrieving a specific contact by ID");
    var options = {filterBy: ["id"],
                   filterOp: "equals",
                   filterValue: sample_id1};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Retrieving a specific contact by givenName");
    var options = {filterBy: ["givenName"],
                   filterOp: "equals",
                   filterValue: properties1.givenName[0]};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, properties1);
      next();
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Modifying contact2");
    if (!findResult1) {
      SpecialPowers.executeSoon(next);
    } else {
      findResult1.impp = properties1.impp = [{value: "phil impp"}];
      req = mozContacts.save(findResult1);
      req.onsuccess = function () {
        var req2 = mozContacts.find(defaultOptions);
        req2.onsuccess = function () {
          is(req2.result.length, 1, "Found exactly 1 contact.");
          findResult1 = req2.result[0];
          ok(findResult1.id == sample_id1, "Same ID");
          checkContacts(findResult1, properties1);
          is(findResult1.impp.length, 1, "Found exactly 1 IMS info.");
          next();
        }
        req2.onerror = onFailure;
      };
      req.onerror = onFailure;
    }
  },
  function () {
    ok(true, "Searching contacts by query");
    var options = {filterBy: ["givenName", "email"],
                   filterOp: "startsWith",
                   filterValue: properties1.givenName[0].substring(0,4)};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Searching contacts by query");
    var options = {filterBy: ["givenName", "email"],
                   filterOp: "startsWith",
                   filterValue: properties1.givenName[0]};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Searching contacts with multiple indices");
    var options = {filterBy: ["email", "givenName"],
                   filterOp: "equals",
                   filterValue: properties1.givenName[1]};
    req = mozContacts.find(options);
    req.onsuccess = function () {
      is(req.result.length, 1, "Found exactly 1 contact.");
      findResult1 = req.result[0];
      ok(findResult1.id == sample_id1, "Same ID");
      checkContacts(findResult1, properties1);
      next();
    };
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Modifying contact3");
    if (!findResult1) {
      SpecialPowers.executeSoon(next);
    } else {
      findResult1.email = [{value: properties1.nickname}];
      findResult1.nickname = ["TEST"];
      var newContact = new mozContact(findResult1);
      req = mozContacts.save(newContact);
      req.onsuccess = function () {
        var options = {filterBy: ["email", "givenName"],
                       filterOp: "startsWith",
                       filterValue: properties1.givenName[0]};
        // One contact has it in nickname and the other in email
        var req2 = mozContacts.find(options);
        req2.onsuccess = function () {
          is(req2.result.length, 2, "Found exactly 2 contacts.");
          ok(req2.result[0].id != req2.result[1].id, "Different ID");
          next();
        }
        req2.onerror = onFailure;
      };
      req.onerror = onFailure;
    }
  },
  function () {
    ok(true, "Deleting contact" + findResult1);
    req = mozContacts.remove(findResult1);
    req.onsuccess = function () {
      var req2 = mozContacts.find(defaultOptions);
      req2.onsuccess = function () {
        is(req2.result.length, 1, "One contact left.");
        findResult1 = req2.result[0];
        next();
      }
      req2.onerror = onFailure;
    }
    req.onerror = onFailure;
  },
  function () {
    ok(true, "Deleting database");
    req = mozContacts.remove(findResult1);
    req.onsuccess =  function () {
      clearTemps();
      next();
    };
    req.onerror = onFailure;
  },
  function() {
    ok(true, "Test JSON.stringify output for mozContact objects");
    var json = JSON.parse(JSON.stringify(new mozContact(properties1)));
    checkContacts(json, properties1);
    next();
  },
  function () {
    ok(true, "all done!\n");
    clearTemps();

    SimpleTest.finish();
  }
];

function next() {
  ok(true, "Begin!");
  if (index >= steps.length) {
    ok(false, "Shouldn't get here!");
    return;
  }
  try {
    var i = index++;
    steps[i]();
  } catch(ex) {
    ok(false, "Caught exception", ex);
  }
}

// Skip tests on Android < 4.0 due to test failures on tbpl (see bugs 897924 & 888891)
if (!isAndroid || androidVersion >= 14) {
  SimpleTest.waitForExplicitFinish();
  addLoadEvent(next);
} else {
  ok(true, "Skip tests on Android < 4.0 (bugs 897924 & 888891");
}

</script>
</pre>
</body>
</html>
